[routeorch] Handle SAI_STATUS_ITEM_NOT_FOUND when setting route entries (#3713)
In dualtor systems, muxorch will program tunnel route entries for mux
neighbors on standby ports. If a route with the same prefix as the mux
tunnel route is learned, a cache entry will be created in routeorch.
Then when the neighbor is moved to an active mux port, the tunnel route
will be deleted.

In certain race-conditions the delete notification will be delayed,
causing a mismatch in routeoch between the internal cache and syncd
state. If a route update is processed during this time, routeorch will
attempt to set the route entry in syncd. This will return an error code
SAI_STATUS_ITEM_NOT_FOUND.

This adds error handling logic to remove the internal cache entry from
routeorch, then attempt to retry the route set.

Signed-off-by: Nikola Dancejic <ndancejic@microsoft.com>