Skip ref counting standby mux neighbor NHs when added to NH group (#3753)
* Fix ref count of neighbors when mux in standby
* Skip ref counting standby mux neighbor NHs when added to NH group (#3753)

What I did
Skip ref counting the MUX neighbor NHs while getting added to NH group when MUX state is in standby.

Why I did it
This issue was causing ref counts of mux NHs to go wrong, observed this in DVS test.
m_neighOrch->getNextHopId(it) this returns tunnel NH id when mux state is standby. However the code was incrementing the ref count of local neighbor NH based on the nh-key.