Fix Orchagent crashes from port down cnt conversion out of range. (#3889)
When a fabric link keep flaky for a long time, Orchagent could crash with the out of range issue. It crashed at the convering the link down count number field from state_db to a uint8_t. This field is a cumulated number of the number of times of a link down, so to prevent out of range issue, it should convert the value to a larger type, such as uint64_t.

The issue is tracked at sonic-net/sonic-buildimage#24020

This PR revisit the related code and convert to related fields to uint64_t.